requirements.txt
numpy
pandas
matplotlib
streamlit
requests
plotly
db-dtypes
google-cloud-bigquery
google-auth
google-cloud
python-dotenv
scipy
sklearn.preprocessing
sklearn.manifold
scikit-learn
seaborn
sklearn.metrics.pairwise
plotly.graph_objects
random
